libs/charts: Fix warnings

+++ libs/charts/iATFTableDlg.cpp
@@ -74,7 +74,7 @@ iATFTableDlg::iATFTableDlg(QWidget* parent, iAChartFunction* func) :
 	connect(table, &QTableWidget::cellChanged, this, &iATFTableDlg::cellValueChanged);
 
 	updateTable();
-	resize(table->columnWidth(0) * columnNames.size(), table->rowHeight(0) * 13);
+	resize(table->columnWidth(0) * static_cast<int>(columnNames.size()), table->rowHeight(0) * 13);
 }
 
 void iATFTableDlg::updateTable()
